Improved synchronization of the audio with the display of letters
Improved sound playback for each letter, now the sound plays once a letter has been displayed.
Before, the sound was played at the beginning of the interpretation of each letter, regardless of how long it took to be displayed, being out of sync with the audio and the display of the letter.